+++ ts/src/program/index.ts
@@ -79,6 +79,7 @@ export class Program<IDL extends Idl = Idl> {
    * });
    * ```
    * @deprecated
+   * Use program.methods.<method>(...args).rpc() instead
    */
   readonly rpc: RpcNamespace<IDL>;
 
@@ -362,7 +363,7 @@ export class Program<IDL extends Idl = Idl> {
    */
   public addEventListener(
     eventName: string,
-    callback: (event: any, slot: number) => void
+    callback: (event: any, slot: number, signature: string) => void
   ): number {
     return this._events.addEventListener(eventName, callback);
   }